--- a/Project.toml
+++ b/Project.toml
@@ -17,7 +17,7 @@ Octavian = "0.2.18, 0.3"
 PrecompileTools = "1"
 TropicalNumbers = "0.2.3, 0.3, 0.4, 0.5"
 VectorizationBase = "0.21"
-julia = "1.6"
+julia = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"