--- a/src/TropicalGEMM.jl
+++ b/src/TropicalGEMM.jl
@@ -7,9 +7,22 @@ using VectorizationBase: contiguous_batch_size, contiguous_axis, val_stride_rank
     vmaximum, fmap, FloatingTypes, IntegerIndex, LazyMulAdd
 using LinearAlgebra: StridedMaybeAdjOrTransMat
 
-export Tropical, TropicalF64, TropicalF32, TropicalF16
+export Tropical, TropicalF64, TropicalF32
 
 include("fallbacks.jl")
 include("gemm.jl")
+
+import PrecompileTools
+PrecompileTools.@setup_workload begin
+    # Putting some things in `@setup_workload` instead of `@compile_workload` can reduce the size of the
+    # precompile file and potentially make loading faster.
+    PrecompileTools.@compile_workload begin
+        for T in (Float32, Float64, Int32, Int64, Int128)
+            x = Tropical.(rand(T, 10, 10))
+            y = Tropical.(rand(T, 10, 10))
+            x * y
+        end
+    end
+end
 
 end