--- a/src/TropicalGEMM.jl
+++ b/src/TropicalGEMM.jl
@@ -5,7 +5,6 @@ using VectorizationBase: OffsetPrecalc, StaticBool, Bit, static, NativeTypes, In
     AbstractMask, NativeTypesExceptBit, AbstractSIMDVector, IndexNoUnroll, AbstractStridedPointer, AbstractSIMD
 using VectorizationBase: contiguous_batch_size, contiguous_axis, val_stride_rank, bytestrides, offsets, memory_reference,
     vmaximum, fmap, FloatingTypes, IntegerIndex, LazyMulAdd
-using LinearAlgebra: StridedMaybeAdjOrTransMat
 
 export Tropical, TropicalF64, TropicalF32
 
@@ -17,10 +16,14 @@ PrecompileTools.@setup_workload begin
     # Putting some things in `@setup_workload` instead of `@compile_workload` can reduce the size of the
     # precompile file and potentially make loading faster.
     PrecompileTools.@compile_workload begin
-        for T in (Float32, Float64, Int32, Int64, Int128)
-            x = Tropical.(rand(T, 10, 10))
-            y = Tropical.(rand(T, 10, 10))
-            x * y
+        for T in (Float32, Float64, Int64)
+            A = Tropical.(rand(T, 10, 10))
+            TA = transpose(A)
+            for x in [A, TA]
+                for y in [A, TA]
+                    x * y
+                end
+            end
         end
     end
 end