--- a/src/gemm.jl
+++ b/src/gemm.jl
@@ -143,7 +143,7 @@ end
 
 # Overwrite the `mul!` in LinearAlgebra (also changes the behavior of `*` in Base)!
 using Octavian
-function LinearAlgebra.mul!(o::StridedMaybeAdjOrTransMat{T}, a::StridedMaybeAdjOrTransMat{T}, b::StridedMaybeAdjOrTransMat{T}, α::Number, β::Number) where {T<:Tropical{<:NativeTypes}}
+function LinearAlgebra.mul!(o::MaybeAdjOrTransMat{T}, a::MaybeAdjOrTransMat{T}, b::MaybeAdjOrTransMat{T}, α::Number, β::Number) where {T<:Tropical{<:NativeTypes}}
     α = _convert_to_static(T, α)
     β = _convert_to_static(T, β)
     Octavian.matmul!(o, a, b, α, β)