## Before
```julia
julia> @time (using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 10, 10)); y = Tropical.(rand(T, 10, 10)); x * y)
 10.316412 seconds (18.54 M allocations: 1.056 GiB, 3.12% gc time, 92.85% compilation time)
```

## After
```julia
julia> @time (using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 10, 10)); y = Tropical.(rand(T, 10, 10)); x * y)
  1.196553 seconds (2.16 M allocations: 182.519 MiB, 5.15% gc time, 1.17% compilation time)
```

This is amazing speed up, which greatly improves the usability of `Octavian` and `TropicalGEMM`.

@chriselrod I notice that in `Octavian`, PrecompileTools is imported but not used. Do you have special concerns like compatibility or binary size? Otherwise, I just confirmed the following speed up of TTFX after the precompilation:
```julia
julia> @time (using Octavian; T = Float64; x = rand(T, 10, 10); y = rand(T, 10, 10); Octavian.matmul(x, y))
  0.796285 seconds (1.38 M allocations: 93.804 MiB, 4.00% gc time, 1.78% compilation time)
```
Before this change, the TTFX is 6s on my machine.

## Update
The binary size is also "amazing":
```bash
-rw-rw-r-- 1 leo leo     19466 Jun 30 14:24 6vpGR_SKm48.ji
-rwxrwxrwx 1 leo leo 486264528 Jun 30 14:24 6vpGR_SKm48.so
```