@chriselrod 2023-06-30
I forgot to remove it in https://github.com/JuliaLinearAlgebra/Octavian.jl/pull/181
Precompilation often results in dynamic dispatches in the precompiled code that goes away if you don't precompile it.