@GiggleLiu 2023-07-01
## Before
```julia
julia> using BenchmarkTools

julia> using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 10, 10)); y = Tropical.(rand(T, 10, 10)); @benchmark $x * $y
BenchmarkTools.Trial: 10000 samples with 900 evaluations.
 Range (min … max):  124.147 ns …  1.115 μs  ┊ GC (min … max): 0.00% … 68.38%
 Time  (median):     132.959 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   150.138 ns ± 99.103 ns  ┊ GC (mean ± σ):  7.41% ±  9.60%

  █▄▂▂▁                                                        ▁
  ██████▇▇▆▆▆▆▆▄▁▁▁█▆▄▃▃▁▁▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▄▃▅▆▄▅▅▄▆▆▅▆ █
  124 ns        Histogram: log(frequency) by time       858 ns <

 Memory estimate: 896 bytes, allocs estimate: 1.

julia> using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 1000, 1000)); y = Tropical.(rand(T, 1000, 1000)); @benchmark $x * $y
BenchmarkTools.Trial: 70 samples with 1 evaluation.
 Range (min … max):  69.017 ms … 85.843 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     70.654 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   71.431 ms ±  2.633 ms  ┊ GC (mean ± σ):  0.17% ± 0.44%

     █                                                         
  ▄▆▄█▇█▃▇▄▆▁▇▄▆▃▄▁▆▆▁▄▃▁▄▁▁▄▁▃▃▁▁▃▁▁▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  69 ms           Histogram: frequency by time        79.8 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 2.
```

## After Precompiling
```julia
julia> using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 10, 10)); y = Tropical.(rand(T, 10, 10)); @benchmark $x * $y
[ Info: Precompiling TropicalGEMM [a4ad3063-64a7-4bad-8738-34ed09bc0236]
BenchmarkTools.Trial: 10000 samples with 888 evaluations.
 Range (min … max):  127.241 ns … 786.229 ns  ┊ GC (min … max): 0.00% … 52.64%
 Time  (median):     134.123 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   148.204 ns ±  61.213 ns  ┊ GC (mean ± σ):  3.81% ±  7.61%

  █▇▄▃▂▂▂▂▂                                                     ▂
  ██████████▇▆▇▆▆▅▅▆▅▅▅▆▅▃▃▄▁▁▅█▆▆▃▄▃▁▃▁▄▃▃▁▃▄▃▄▁▅▅▅▄▅▄▅▆▅▄▅▅▄▅ █
  127 ns        Histogram: log(frequency) by time        557 ns <

 Memory estimate: 896 bytes, allocs estimate: 1.

julia> using TropicalGEMM; T = Float64; x = Tropical.(rand(T, 1000, 1000)); y = Tropical.(rand(T, 1000, 1000)); @benchmark $x * $y
BenchmarkTools.Trial: 72 samples with 1 evaluation.
 Range (min … max):  68.306 ms … 73.864 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     68.889 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   69.528 ms ±  1.179 ms  ┊ GC (mean ± σ):  0.17% ± 0.44%

   ▃ █▅ ▂                                                      
  ▄█▄████▇▄▇▄▁▁▄▁▄▄▄▁▄▄▄▁▁▅▄▅▁▄▄▅▄▅▄▄▅▅▄▄▄▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▄ ▁
  68.3 ms         Histogram: frequency by time        72.7 ms <

 Memory estimate: 7.63 MiB, allocs estimate: 2.
```