@GiggleLiu 2023-07-01
I used your script for benchmarking Octavian, but did not see any dispatch issue.

## Without precompile,
```julia
julia> using BenchmarkTools

julia> using Octavian
[ Info: Precompiling Octavian [6fd5a793-0b7e-452c-907f-f8bfe9c57db4]
A^[[A
julia> A = rand(-1_000:1_000, 200, 200);

julia> B = rand(-1_000:1_000, 200, 200);

julia> C = similar(A);

julia> Af64 = Float64.(A); Bf64 = Float64.(B); Cf64 = similar(Af64);

julia> @benchmark matmul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  278.507 μs … 524.333 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     286.815 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   289.723 μs ±  14.144 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃▃   ▅█▅▄▃▃▂▂▁▁                                               ▂
  ██▆▃▄███████████▇█▇▇▇▆▆█████▇▆▆▅▆▆▆▆▆▆▆▆▆▆▆▇▅▅▆▆▅▅▆▆▄▄▄▃▅▅▁▃▅ █
  279 μs        Histogram: log(frequency) by time        353 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

## With precompile

```julia
julia> using Octavian
[ Info: Precompiling Octavian [6fd5a793-0b7e-452c-907f-f8bfe9c57db4]

julia> A = rand(-1_000:1_000, 200, 200);

julia> B = rand(-1_000:1_000, 200, 200);

julia> C = similar(A);

julia> using BenchmarkTools

julia> Af64 = Float64.(A); Bf64 = Float64.(B); Cf64 = similar(Af64);

julia> @benchmark matmul!($Cf64, $Af64, $Bf64)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  283.227 μs … 553.582 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     285.139 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   286.645 μs ±  12.436 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅▅█▇█▆▅▄▄▃▃▃▃▂▂▂▁▁▁ ▁▁                                        ▂
  ███████████████████████▇██▇▆▇▇▇▇▆▆▅▅▆▅▄▄▅▄▃▄▅▃▃▄▅▅▃▅▅▃▄▃▁▄▄▅▃ █
  283 μs        Histogram: log(frequency) by time        309 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

## Precompiling script
```julia
@setup_workload begin
    # Putting some things in `@setup_workload` instead of `@compile_workload` can reduce the size of the
    # precompile file and potentially make loading faster.
    @compile_workload begin
        for T in (Float32, Float64)
            x = rand(T, 10, 10)
            y = rand(T, 10, 10)
            Octavian.matmul(x, y)
        end
    end
end
```